(chore): move owner check to pkg/e2e
